Made Sprint GC stop playing when local player stops sprinting (vs waiting for Server)
Previously when the local player stopped sprinting, the sprint GC would continue playing until the sprint GE was removed by the server since we can't predictively remove GEs.

By removing the sprint GC from the GE and tying it to the lifetime of the ability with the GA execute GC node, the local player's sprint GA will end before the server's making the GC stop playing when the local player's GA ends.