Fixed projectile collision in a standalone game
In a standalone game, if the player was in certain locations the projectiles didn't collide. Enabling CCD on the sphere mesh collision didn't help. Reducing the speed of projectiles seems to fix the bug, but I want my projectiles at 7000 u/s.

This issue isn't present in UE4 4.24 which leads me to believe it might have been a 4.23 bug (it's unclear). I swapped from using the static mesh as the collision volume to a basic sphere collision component. This seems to have fixed the issue in 4.23 while keeping the projectile speed at 7000 u/s. I'm not really sure why a Sphere Component is better than the Sphere Static Mesh, but I wager the Sphere Component is more efficient for collision anyway so it's a better solution overall.