Swapped RefreshAbilityActorInfo for InitAbilityActorInfo
Refresh should be used when updating things like AnimInstance, CharacterMovementComponent, etc in the ActorInfo.

> AvatarActor will be the same but we will look for new AnimInstance, MovementComponent, PlayerController, etc.

When we're possessing on the client, we're setting a new AvatarActor and therefor we really should be calling `InitAbilityActor()`. While `RefreshAbilityActorInfo()` seems to work okay, I don't know what unintended consequences may happen in more complex projects, especially when possessing different classes. It's best to use the functions as intended.