Use correct activation info for ending jump
An ability's `CurrentSpecHandle`, `CurrentActorInfo` and `CurrentActivationInfo` are not set when the abiilty is non-instanced.
Passing them to `EndAbility` will make the call silently fail (because `IsEndAbilityValid` fails to get a pointer to the ASC from the `ActorInfo`), so the ability never ends.
The callsite in the jump ability was likely never called so it didn't cause any issues, but it's better to be consistent in case someone uses it as reference.