Fix shared Cooldown GE bug with tags writing to GA CDO
MutableTags writes to the TempCooldownTags variable on the GA CDO. Make sure to clear the container before rebuilding it. This prevents issues where Design might swap abilities from one ability slot to another, changing the cooldown tags in the process. This would have the ability tags from the previous slot still living in the GA's CDO and put both slots on cooldown when the GA is activated.