Fixed crash in packaged build from DamageNumberClass not being packaged
It crashed because the DamageNumber class wasn't being packaged since there wasn't a hard reference to it in any asset that was being packaged. We could have fixed it by including its directory in the list of directories to always package in the Project Settings. But, static loading assets is not a good idea.

Removed the call to static load and assigned the class in the BP_PlayerController's class defaults instead. Now there's a hard reference to the widget class so it will be packaged by default.